# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Page Cloner API
  version: 1.0.0
  summary: Asynchronous web page cloning, sanitization, and rebundling service.
  description: |
    The Page Cloner API renders any public URL with a headless browser, strips
    scripts/tracking, exposes detected forms and links for editing, and emits
    a downloadable HTML or ZIP bundle.

    ## Lifecycle
    1. `POST /v1/clones` to enqueue a clone job (async, returns `202`).
    2. Poll `GET /v1/clones/{id}` or receive a `webhook_url` callback when `status = ready`.
    3. Inspect detected `forms` and `links`; mutate via `PATCH` or bulk endpoints.
    4. `POST /v1/clones/{id}/build` to materialize a final artifact.
    5. `GET .../builds/{buildId}/download` to stream the file.
    6. `DELETE /v1/clones/{id}` purges the job and all associated assets.

    ## Conventions
    - All resource IDs are **ULIDs** (Crockford Base32, 26 chars) — sortable by creation time.
    - Errors follow **RFC 7807** (`application/problem+json`).
    - List endpoints use **cursor pagination** (`limit`, `cursor`).
    - GET responses on stable resources include strong `ETag`s; clients should send
      `If-None-Match` to save bandwidth and `If-Match` on PATCH for optimistic concurrency.
    - All timestamps are RFC 3339 in UTC.
  contact:
    name: Page Cloner Platform
    email: api@pagecloner.example
  license:
    name: Proprietary
    url: https://pagecloner.example/terms
servers:
  - url: https://api.pagecloner.example
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/clones/{id}/build:
    parameters:
      - $ref: '#/components/parameters/CloneId'
    post:
      tags: [Builds]
      operationId: createBuild
      summary: Apply edits and produce a final bundle
      description: |
        Snapshots the current edits and starts a build job. Returns `202` with
        a `Location` pointing at the build status resource.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildOptions'
            examples:
              singleHtml:
                value:
                  format: html
                  include_assets: true
              zipBundle:
                value:
                  format: zip
                  include_assets: true
      responses:
        '202':
          description: Build job accepted.
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildJob'
              examples:
                queued:
                  value:
                    id: bld_01HW3KB9C0D1E2F3G4H5J6K7L8
                    clone_id: 01HW3K9P2N7XQK4D6V5RYMC1ZB
                    status: queued
                    options:
                      format: zip
                      include_assets: true
                    created_at: '2026-05-02T14:25:00Z'
                    updated_at: '2026-05-02T14:25:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Clone is not in a buildable state (must be `ready`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/clones/{id}/builds/{buildId}:
    parameters:
      - $ref: '#/components/parameters/CloneId'
      - $ref: '#/components/parameters/BuildId'
    get:
      tags: [Builds]
      operationId: getBuild
      summary: Get build job status
      description: |
        Returns the build's status. When `status = ready`, the payload includes
        a time-limited `download_url` (and the canonical `download` API path).
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Build state.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildJob'
              examples:
                ready:
                  value:
                    id: bld_01HW3KB9C0D1E2F3G4H5J6K7L8
                    clone_id: 01HW3K9P2N7XQK4D6V5RYMC1ZB
                    status: ready
                    options:
                      format: zip
                      include_assets: true
                    artifact:
                      bytes: 1923004
                      content_type: application/zip
                      sha256: 3a1f5d9b2c87e0f4a6d8b2e4c6a8d0f2e4c6a8d0f2e4c6a8d0f2e4c6a8d0f2e4
                      filename: example-com-landing.zip
                    download_url: https://cdn.pagecloner.example/builds/bld_01HW3KB9C0D1E2F3G4H5J6K7L8/example-com-landing.zip?token=...
                    download_expires_at: '2026-05-02T15:25:00Z'
                    created_at: '2026-05-02T14:25:00Z'
                    updated_at: '2026-05-02T14:25:14Z'
                    links:
                      self: /v1/clones/01HW3K9P2N7XQK4D6V5RYMC1ZB/builds/bld_01HW3KB9C0D1E2F3G4H5J6K7L8
                      download: /v1/clones/01HW3K9P2N7XQK4D6V5RYMC1ZB/builds/bld_01HW3KB9C0D1E2F3G4H5J6K7L8/download
        '304':
          description: Not Modified.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/clones/{id}/builds/{buildId}/download:
    parameters:
      - $ref: '#/components/parameters/CloneId'
      - $ref: '#/components/parameters/BuildId'
    get:
      tags: [Builds]
      operationId: downloadBuild
      summary: Download the final bundle
      description: |
        Streams the artifact. `Content-Type` is `text/html` for `format=html`
        bundles and `application/zip` for ZIPs. Returns `409` if the build
        is not yet `ready` and `410 Gone` if the artifact has been purged.
      responses:
        '200':
          description: Artifact stream.
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="example-com-landing.zip"'
            ETag:
              $ref: '#/components/headers/ETag'
            X-Content-Sha256:
              schema:
                type: string
                description: Hex-encoded SHA-256 of the artifact.
          content:
            text/html:
              schema:
                type: string
                contentMediaType: text/html
            application/zip:
              schema:
                type: string
                format: binary
                contentMediaType: application/zip
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Build not yet `ready`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '410':
          description: Artifact purged or expired.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OAuth 2.0 Bearer access token (JWT). Use for end-user / first-party
        clients. Tokens carry `scope` claims; minimum required scope is
        `clones:write` for mutations and `clones:read` for GETs.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Long-lived API key for server-to-server integrations. Treat as a secret;
        rotate via the dashboard.
  parameters:
    CloneId:
      name: id
      in: path
      required: true
      description: Clone job ID (ULID).
      schema:
        $ref: '#/components/schemas/UlidId'
...
    BuildId:
      name: buildId
      in: path
      required: true
      description: Build job ID (prefixed ULID, e.g. `bld_…`).
      schema:
        $ref: '#/components/schemas/PrefixedUlid'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Client-generated unique string (UUIDv4 or ULID, max 128 chars).
        Replays within 24 hours return the original response.
      schema:
        type: string
        maxLength: 128
        pattern: '^[A-Za-z0-9_\-]+$'
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: ETag value to enable `304 Not Modified` short-circuit.
      schema:
        type: string
  headers:
    ETag:
      description: Strong entity tag for the resource representation.
      schema:
        type: string
        example: '"W3v1k2"'
...
  schemas:
    UlidId:
      type: string
      description: Crockford Base32 ULID (26 chars), sortable by creation time.
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
      example: 01HW3K9P2N7XQK4D6V5RYMC1ZB
    PrefixedUlid:
      type: string
      description: Resource-prefixed ULID (e.g. `frm_…`, `lnk_…`, `bld_…`).
      pattern: '^(frm|lnk|bld)_[0-9A-HJKMNP-TV-Z]{26}$'
      example: frm_01HW3KA8H8Y4N5Q1QH8WV2GZJ7
...
    BuildStatus:
      type: string
      enum:
        - queued
        - building
        - ready
        - failed
      description: Lifecycle state of a build job.
...
    JobError:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
            - render_timeout
            - dns_failure
            - tls_error
            - http_error
            - blocked_by_robots
            - too_large
            - sanitization_failed
            - internal_error
        message:
          type: string
        http_status:
          type: integer
          description: Upstream HTTP status, if applicable.
...
    BuildOptions:
      type: object
      required: [format]
      additionalProperties: false
      properties:
        format:
          type: string
          enum: [html, zip]
        include_assets:
          type: boolean
          default: true
          description: |
            When `true`, downloads and bundles referenced images/CSS/fonts.
            For `format=html`, assets are inlined as data-URIs.
            For `format=zip`, assets are placed in `/assets/` and rewritten.
    BuildArtifact:
      type: object
      required: [bytes, content_type, sha256, filename]
      additionalProperties: false
      properties:
        bytes:
          type: integer
          minimum: 0
        content_type:
          type: string
          example: application/zip
        sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
        filename:
          type: string
    BuildJob:
      type: object
      required: [id, clone_id, status, options, created_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/PrefixedUlid'
        clone_id:
          $ref: '#/components/schemas/UlidId'
        status:
          $ref: '#/components/schemas/BuildStatus'
        options:
          $ref: '#/components/schemas/BuildOptions'
        artifact:
          $ref: '#/components/schemas/BuildArtifact'
        download_url:
          type: string
          format: uri
          description: |
            Time-limited, pre-signed URL for direct CDN download. Only present
            when `status = ready`.
        download_expires_at:
          type: string
          format: date-time
        error:
          $ref: '#/components/schemas/JobError'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        links:
          type: object
          additionalProperties: false
          properties:
            self:
              type: string
              format: uri-reference
            download:
              type: string
              format: uri-reference
...
    Problem:
      type: object
      description: RFC 7807 Problem Details object.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
          description: A URI reference identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying the specific occurrence.
        code:
          type: string
          description: Machine-readable application error code.
        errors:
          type: array
          description: |
            Field-level validation errors (only on `422`). Each entry points
            at a JSON Pointer path within the request body.
          items:
            type: object
            required: [pointer, message]
            properties:
              pointer:
                type: string
                example: /options/timeout_ms
              message:
                type: string
                example: must be <= 60000
              code:
                type: string
                example: out_of_range
  responses:
    BadRequest:
      description: Malformed request (invalid JSON, missing required fields).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://pagecloner.example/problems/bad-request
            title: Bad Request
            status: 400
            detail: Request body is not valid JSON.
            code: bad_request
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://pagecloner.example/problems/unauthorized
            title: Unauthorized
            status: 401
            detail: Bearer token is missing or expired.
            code: unauthorized
    NotFound:
      description: Resource does not exist or has been purged.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://pagecloner.example/problems/not-found
            title: Not Found
            status: 404
            detail: No clone job with id 01HW3K9P2N7XQK4D6V5RYMC1ZB.
            code: not_found
...
    UnprocessableEntity:
      description: Request was syntactically valid but failed semantic validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://pagecloner.example/problems/validation
            title: Unprocessable Entity
            status: 422
            detail: One or more fields failed validation.
            code: validation_failed
            errors:
              - pointer: /url
                message: must be a valid http(s) URL
                code: invalid_format
              - pointer: /options/timeout_ms
                message: must be <= 60000
                code: out_of_range
...